Renaming api fields component file
